ddasd

+++ src/main/java/com/excel/services/ExcelServices.java
@@ -1,6 +1,9 @@
 package com.excel.services;
 
 public interface ExcelServices {
-	public String readExcelColumn(Long row, String column, String filePath);
-	public void writeExcelColumn(Long row, String column, String value, String filePath);
+	
+	public String readExcelColumn(int row, String column, String filePath);
+	public void writeExcelColumn(int row, String column, String value, String filePath);
+
+
 }
